@@ -21,23 +21,28 @@
     "@emotion/react": "11.7.1",
     "@emotion/styled": "11.6.0",
     "@mui/icons-material": "5.4.1",
-    "@mui/lab": "^5.0.0-alpha.99",
     "@mui/material": "5.4.1",
     "@mui/styled-engine": "5.4.1",
-    "@mui/x-date-pickers": "^5.0.0",
     "@testing-library/jest-dom": "5.16.2",
     "@testing-library/react": "12.1.2",
     "@testing-library/user-event": "13.5.0",
+    "antd": "^4.23.4",
     "axios": "^0.27.2",
     "chart.js": "3.4.1",
     "chroma-js": "2.4.2",
+    "date-fns": "^2.29.3",
+    "dayjs": "^1.11.5",
     "formik": "^2.2.9",
+    "google-map-react": "^2.2.0",
     "prop-types": "15.8.1",
-    "react": "17.0.2",
+    "react": "^18.2.0",
     "react-chartjs-2": "3.0.4",
-    "react-dom": "17.0.2",
+    "react-dom": "^18.2.0",
     "react-github-btn": "1.2.1",
+    "react-images-viewer": "^1.7.1",
+    "react-loader-spinner": "^5.3.4",
     "react-material-file-upload": "^0.0.4",
+    "react-material-ui-carousel": "^3.4.2",
     "react-router-dom": "6.2.1",
     "react-scripts": "5.0.0",
     "react-table": "7.7.0",
