@@ -22,38 +22,40 @@ import MDTypography from "components/MDTypography";
 import MDInput from "components/MDInput";
 import MDButton from "components/MDButton";
 
-import { useFormik } from "formik";
 // Authentication layout components
 import CoverLayout from "layouts/authentication/components/CoverLayout";
-
+useState
 // Images
 import bgImage from "assets/images/bg-reset-cover.jpeg";
+import { useFormik } from "formik";
 import { useState } from "react";
-import { userVerifyEmail } from "validations/Uservalidation";
-import { Typography } from "@mui/material";
-import { getByEmail } from "api";
 import { useNavigate } from "react-router-dom";
+import { verifyOtp } from "validations/Uservalidation";
+import { verifyAdminOtp } from "api";
+import { Typography } from "@mui/material";
 
 
 function Cover() {
+
   const [error, setError] = useState("");
   const navigate = useNavigate();
+  const adminId = localStorage.getItem("adminId")
   const initialValues = {
 
-    email: "",
+    otp: "",
+    adminId: adminId
 
   }
   const { values, errors, handleChange, handleSubmit } = useFormik({
     initialValues: initialValues,
-    validationSchema: userVerifyEmail,
+    validationSchema: verifyOtp,
     onSubmit: async (values) => {
       try {
-        localStorage.removeItem("adminId");
-        const { data } = await getByEmail(values);
+        const { data } = await verifyAdminOtp(values);
         console.log(data);
-        localStorage.setItem("adminId", data.admin._id);
+
         if (data.status == 200) {
-          navigate("/otp")
+          navigate("/newPassword")
         }
       }
       catch (e) {
@@ -65,6 +67,7 @@ function Cover() {
       }
     }
   })
+
 
   return (
     <CoverLayout coverHeight="50vh" image={bgImage}>
@@ -81,25 +84,25 @@ function Cover() {
           textAlign="center"
         >
           <MDTypography variant="h3" fontWeight="medium" color="white" mt={1}>
-            Forgot Password
+            Enter OTP
           </MDTypography>
           <MDTypography display="block" variant="button" color="white" my={1}>
-            You will receive an e-mail in maximum 60 seconds
+            Check your Emails OTP will be there
           </MDTypography>
         </MDBox>
         <MDBox pt={4} pb={3} px={3}>
           <MDBox component="form" role="form">
             <MDBox mb={4}>
-              <MDInput type="email" label="Email" variant="standard" name="email" value={values.email} onChange={handleChange} fullWidth />
+              <MDInput type="number" label="OTP" value={values.otp} variant="standard" name="otp" onChange={handleChange} fullWidth />
               {<Typography display="block" variant="string" color="red" sx={{ fontSize: "12px" }} my={1}>
-                {errors.email}
+                {errors.otp}
               </Typography>}
             </MDBox>
             <MDBox mt={6} mb={1}>
               <MDButton variant="gradient" color="info" fullWidth onClick={handleSubmit}>
                 submit
               </MDButton>
-              {<Typography display="block" variant="string" color="red" sx={{ fontSize: "12px", textAlign: "center" }} my={1}>
+              {<Typography display="block" variant="string" color="red" sx={{ fontSize: "12px" }} my={1}>
                 {error}
               </Typography>}
             </MDBox>
